Update random usage of different integer sizes to use a consistent typedef

+++ crates/engine/src/moldudp64/engine.rs
@@ -283,8 +283,8 @@ mod tests {
                     side: OrderSide::Ask,
                     kind: OrderType::Update {
                         old_id: i - 1,
-                        qty: i,
-                        price: i,
+                        qty: i as u32,
+                        price: i as u32,
                     },
                     extra: L3EventExtra::None,
                 }),
@@ -300,8 +300,8 @@ mod tests {
             let trade_event = MarketEvent {
                 id: 0,
                 kind: MarketEventType::Trade(TradeEvent {
-                    price: i,
-                    quantity: i,
+                    price: i as u32,
+                    quantity: i as u32,
                     aggressor_side: OrderSide::Ask,
                     maker_id: i,
                 }),
